Refactor of models
