+++ src/models/models/__init__.py
@@ -1,0 +1,2 @@
+from src.models.models.models import *
+from src.models.models.bases import *
